Merge from Master

+++ app/Api/V1/Controllers/UserController.php
@@ -2,47 +2,44 @@
 
 /**
  * UserController.php
- * Copyright (c) 2018 thegrumpydictator@gmail.com
+ * Copyright (c) 2019 thegrumpydictator@gmail.com
  *
- * This file is part of Firefly III.
+ * This file is part of Firefly III (https://github.com/firefly-iii).
  *
- * Firefly III is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
  *
- * Firefly III is distributed in the hope that it will be useful,
+ * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
+ * GNU Affero General Public License for more details.
  *
- * You should have received a copy of the GNU General Public License
- * along with Firefly III. If not, see <http://www.gnu.org/licenses/>.
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
 declare(strict_types=1);
 
 namespace FireflyIII\Api\V1\Controllers;
 
-use FireflyIII\Api\V1\Requests\UserRequest;
+use FireflyIII\Api\V1\Requests\UserStoreRequest;
+use FireflyIII\Api\V1\Requests\UserUpdateRequest;
 use FireflyIII\Exceptions\FireflyException;
 use FireflyIII\Repositories\User\UserRepositoryInterface;
 use FireflyIII\Transformers\UserTransformer;
 use FireflyIII\User;
 use Illuminate\Http\JsonResponse;
-use Illuminate\Http\Request;
 use Illuminate\Pagination\LengthAwarePaginator;
-use League\Fractal\Manager;
 use League\Fractal\Pagination\IlluminatePaginatorAdapter;
 use League\Fractal\Resource\Collection as FractalCollection;
 use League\Fractal\Resource\Item;
-use League\Fractal\Serializer\JsonApiSerializer;
 
 
 /**
  * Class UserController.
  *
- * @SuppressWarnings(PHPMD.CouplingBetweenObjects)
  */
 class UserController extends Controller
 {
@@ -52,6 +49,8 @@ class UserController extends Controller
 
     /**
      * UserController constructor.
+     *
+     * @codeCoverageIgnore
      */
     public function __construct()
     {
@@ -69,39 +68,35 @@ class UserController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param  \FireflyIII\User $user
+     * @param User $user
      *
      * @return JsonResponse
      * @throws FireflyException
+     * @codeCoverageIgnore
      */
     public function delete(User $user): JsonResponse
     {
         /** @var User $admin */
         $admin = auth()->user();
-        if ($this->repository->hasRole($admin, 'owner')) {
+        if ($admin->id !== $user->id && $this->repository->hasRole($admin, 'owner')) {
             $this->repository->destroy($user);
 
             return response()->json([], 204);
         }
-        throw new FireflyException('No access to method.'); // @codeCoverageIgnore
+        throw new FireflyException('200025: No access to function.'); // @codeCoverageIgnore
     }
 
     /**
      * Display a listing of the resource.
      *
-     * @param Request $request
-     *
      * @return JsonResponse
+     * @codeCoverageIgnore
      */
-    public function index(Request $request): JsonResponse
+    public function index(): JsonResponse
     {
         // user preferences
         $pageSize = (int)app('preferences')->getForUser(auth()->user(), 'listPageSize', 50)->data;
-
-        // make manager
-        $manager = new Manager();
-        $baseUrl = $request->getSchemeAndHttpHost() . '/api/v1';
-        $manager->setSerializer(new JsonApiSerializer($baseUrl));
+        $manager  = $this->getManager();
 
         // build collection
         $collection = $this->repository->all();
@@ -113,7 +108,11 @@ class UserController extends Controller
         $paginator->setPath(route('api.v1.users.index') . $this->buildParams());
 
         // make resource
-        $resource = new FractalCollection($users, new UserTransformer($this->parameters), 'users');
+        /** @var UserTransformer $transformer */
+        $transformer = app(UserTransformer::class);
+        $transformer->setParameters($this->parameters);
+
+        $resource = new FractalCollection($users, $transformer, 'users');
         $resource->setPaginator(new IlluminatePaginatorAdapter($paginator));
 
         return response()->json($manager->createData($resource)->toArray())->header('Content-Type', 'application/vnd.api+json');
@@ -122,24 +121,21 @@ class UserController extends Controller
     /**
      * Show a single user.
      *
-     * @param Request $request
-     * @param User    $user
+     * @param User $user
      *
      * @return JsonResponse
+     * @codeCoverageIgnore
      */
-    public function show(Request $request, User $user): JsonResponse
+    public function show(User $user): JsonResponse
     {
         // make manager
-        $manager = new Manager();
-        $baseUrl = $request->getSchemeAndHttpHost() . '/api/v1';
-        $manager->setSerializer(new JsonApiSerializer($baseUrl));
-
-        // add include parameter:
-        $include = $request->get('include') ?? '';
-        $manager->parseIncludes($include);
-
+        $manager = $this->getManager();
         // make resource
-        $resource = new Item($user, new UserTransformer($this->parameters), 'users');
+        /** @var UserTransformer $transformer */
+        $transformer = app(UserTransformer::class);
+        $transformer->setParameters($this->parameters);
+
+        $resource = new Item($user, $transformer, 'users');
 
         return response()->json($manager->createData($resource)->toArray())->header('Content-Type', 'application/vnd.api+json');
     }
@@ -147,26 +143,23 @@ class UserController extends Controller
     /**
      * Store a new user.
      *
-     * @param UserRequest $request
+     * @param UserStoreRequest $request
      *
      * @return JsonResponse
      */
-    public function store(UserRequest $request): JsonResponse
+    public function store(UserStoreRequest $request): JsonResponse
     {
-        $data = $request->getAll();
-        $user = $this->repository->store($data);
-
-        // make manager
-        $manager = new Manager();
-        $baseUrl = $request->getSchemeAndHttpHost() . '/api/v1';
-        $manager->setSerializer(new JsonApiSerializer($baseUrl));
-
-        // add include parameter:
-        $include = $request->get('include') ?? '';
-        $manager->parseIncludes($include);
+        $data    = $request->getAll();
+        $user    = $this->repository->store($data);
+        $manager = $this->getManager();
 
         // make resource
-        $resource = new Item($user, new UserTransformer($this->parameters), 'users');
+
+        /** @var UserTransformer $transformer */
+        $transformer = app(UserTransformer::class);
+        $transformer->setParameters($this->parameters);
+
+        $resource = new Item($user, $transformer, 'users');
 
         return response()->json($manager->createData($resource)->toArray())->header('Content-Type', 'application/vnd.api+json');
     }
@@ -174,27 +167,22 @@ class UserController extends Controller
     /**
      * Update a user.
      *
-     * @param UserRequest $request
-     * @param User        $user
+     * @param UserUpdateRequest $request
+     * @param User              $user
      *
      * @return JsonResponse
      */
-    public function update(UserRequest $request, User $user): JsonResponse
+    public function update(UserUpdateRequest $request, User $user): JsonResponse
     {
-        $data = $request->getAll();
-        $user = $this->repository->update($user, $data);
-
-        // make manager
-        $manager = new Manager();
-        $baseUrl = $request->getSchemeAndHttpHost() . '/api/v1';
-        $manager->setSerializer(new JsonApiSerializer($baseUrl));
-
-        // add include parameter:
-        $include = $request->get('include') ?? '';
-        $manager->parseIncludes($include);
-
+        $data    = $request->getAll();
+        $user    = $this->repository->update($user, $data);
+        $manager = $this->getManager();
         // make resource
-        $resource = new Item($user, new UserTransformer($this->parameters), 'users');
+        /** @var UserTransformer $transformer */
+        $transformer = app(UserTransformer::class);
+        $transformer->setParameters($this->parameters);
+
+        $resource = new Item($user, $transformer, 'users');
 
         return response()->json($manager->createData($resource)->toArray())->header('Content-Type', 'application/vnd.api+json');
